💡 chore    : 요청된 뷰,시리얼라이져, 파일 dropdown_list로 통합
